requirements.txt
streamlit>=1.32.0
upstash-vector>=0.8.0
langchain>=0.3.0
langchain-aws>=0.2.15
langchain-community>=0.3.20
langchain-openai>=0.3.9
opensearch-py>=2.8.0
boto3>=1.34.0
faiss-cpu>=1.7.4
pytest>=8.0.0
pytest-asyncio>=0.21.1
logging-utils>=0.3.0
typing-extensions>=4.5.0
pydantic>=2.0.0
pydantic-settings>=2.1.0
tenacity>=8.2.2
numpy>=1.24.3
pandas>=2.0.3
watchdog>=6.0.0
#PyYAML > 5.3 fails to build
python-dotenv==0.21.1
PyYAML==5.3
